refactor: modify Benchmarker part for sunset petsard util
