+++ petsard/loader/benchmarker.py
@@ -1,3 +1,4 @@
+import hashlib
 import logging
 import os
 from abc import ABC, abstractmethod
@@ -5,7 +6,23 @@ from abc import ABC, abstractmethod
 import requests
 
 from petsard.exceptions import BenchmarkDatasetsError
-from petsard.util import digest_sha256
+
+
+def digest_sha256(filepath):
+    """
+    Calculate SHA-256 value of file. Load 128KB at one time.
+    ...
+    Args:
+        filepath (str) Openable file full path.
+    ...
+    return:
+        (str) SHA-256 value of file.
+    """
+    sha256hash = hashlib.sha256()
+    with open(filepath, "rb") as f:
+        for byte_block in iter(lambda: f.read(131072), b""):
+            sha256hash.update(byte_block)
+    return sha256hash.hexdigest()
 
 
 class BaseBenchmarker(ABC):
@@ -85,13 +102,13 @@ class BaseBenchmarker(ABC):
                         f"Downloaded file SHA-256 mismatch: {self.config['benchmark_filename']} from "
                         f"{self.config['benchmark_bucket_name']}. "
                     )
-                except OSError:
+                except OSError as e:
                     self._logger.error(
                         f"Failed to remove file: {self.config['filepath']}. Please delete it manually."
                     )
                     raise OSError(
                         f"Failed to remove file: {self.config['filepath']}. Please delete it manually."
-                    )
+                    ) from e
 
 
 class BenchmarkerRequests(BaseBenchmarker):
